Refactored and Tidied up. The last commit missed the refactoring, though it claimed one
