--- acceptance-tests/sample.crosschaincoordination.properties
@@ -1,3 +1,0 @@
-#Sample code properties file
-#Mon Nov 18 14:27:56 AEST 2019
-privateKey=29949a9f534dd065a31decf686a388b9da0a0fe35b5165ac10582d74fe2893e9
