+++ acceptance-tests/src/test/java/org/hyperledger/besu/tests/acceptance/crosschain/lockability/SimpleIsLockableCrosschain.sol
@@ -1,0 +1,24 @@
+/*
+ * Copyright 2018 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+pragma solidity ^0.5.0;
+import "../common/Crosschain.sol";
+
+// Simple contract for use with IsLockable tests
+contract SimpleIsLockableCrosschain is Crosschain {
+    uint256 public value;
+
+    function setValue(uint256 _val) external {
+        value = _val;
+    }
+
+}
